Improve and better document the source layout of the planner.
NB: This patch only moves code around to different files, changes
comments and adds non-functional interfaces (= no logic changes).

The "planner" object in the sql package is a bit a showcase of the
"God class" antipattern. This patch alleviates this situation by
re-organizing the services of the planner in separate files that
highlight its separate roles:

- its role as manager of DB descriptors, via a DescriptorAccessor
  interface which documents all the related methods (in
  descriptor.go).

- its role as accessor for database descriptors via a DatabaseAccessor
  interface which documents all the related methods (in database.go).

- its role as accessor for table descriptors and table schema updates,
  via a SchemaAccessor interface which documents all the related
  methods (in table.go).

The planner constructor and struct definition is moved away from
`plan.go` into a new `planner.go`. `plan.go` remains to define the
planNode interface and assert its implementation by the various
concrete structs.

The methods corresponding to each sub-interface are also moved to the
primary source file for that interface.
