+++ sql/table.go
@@ -26,7 +26,9 @@ import (
 	"github.com/cockroachdb/cockroach/roachpb"
 	"github.com/cockroachdb/cockroach/sql/parser"
 	"github.com/cockroachdb/cockroach/sql/sqlbase"
+	"github.com/cockroachdb/cockroach/util"
 	"github.com/cockroachdb/cockroach/util/encoding"
+	"github.com/cockroachdb/cockroach/util/log"
 )
 
 var testDisableTableLeases bool
@@ -58,10 +60,97 @@ func (tk tableKey) Name() string {
 	return tk.name
 }
 
-// getTableDesc returns a table descriptor, or nil if the descriptor is not
-// found.
-// If you want to transform the not found condition into an error, use
-// tableDoesNotExistError().
+var errDescriptorNotFound = errors.New("descriptor not found")
+
+// getTableDescFromID retrieves the table descriptor for the table
+// ID passed in using an existing txn. Teturns an error if the
+// descriptor doesn't exist or if it exists and is not a table.
+func getTableDescFromID(txn *client.Txn, id sqlbase.ID) (*sqlbase.TableDescriptor, error) {
+	desc := &sqlbase.Descriptor{}
+	descKey := sqlbase.MakeDescMetadataKey(id)
+
+	if err := txn.GetProto(descKey, desc); err != nil {
+		return nil, err
+	}
+	table := desc.GetTable()
+	if table == nil {
+		return nil, errDescriptorNotFound
+	}
+	return table, nil
+}
+
+// getKeysForTableDescriptor retrieves the KV keys corresponding
+// to the zone, name and descriptor of a table.
+func getKeysForTableDescriptor(
+	tableDesc *sqlbase.TableDescriptor,
+) (zoneKey roachpb.Key, nameKey roachpb.Key, descKey roachpb.Key) {
+	zoneKey = sqlbase.MakeZoneKey(tableDesc.ID)
+	nameKey = sqlbase.MakeNameMetadataKey(tableDesc.ParentID, tableDesc.GetName())
+	descKey = sqlbase.MakeDescMetadataKey(tableDesc.ID)
+	return
+}
+
+func updateCommitBy(commitBy time.Time, lease *LeaseState) time.Time {
+	if commitBy.IsZero() || commitBy.After(lease.Expiration()) {
+		return lease.Expiration()
+	}
+	return commitBy
+}
+
+// SchemaAccessor provides helper methods for using the SQL schema.
+type SchemaAccessor interface {
+	// getTableNames retrieves the list of qualified names of tables
+	// present in the given database.
+	getTableNames(dbDesc *sqlbase.DatabaseDescriptor) (parser.QualifiedNames, error)
+
+	// expandTableGlob expands wildcards from the end of `expr` and
+	// returns the list of matching tables.
+	// `expr` is possibly modified to be qualified with the database it refers to.
+	// `expr` is assumed to be of one of several forms:
+	// 		database.table
+	// 		table
+	// 		*
+	expandTableGlob(expr *parser.QualifiedName) (parser.QualifiedNames, error)
+
+	// getTableDesc returns a table descriptor, or nil if the descriptor
+	// is not found.  If you want to transform the not found condition
+	// into an error, use tableDoesNotExistError().
+	getTableDesc(qname *parser.QualifiedName) (*sqlbase.TableDescriptor, error)
+
+	// getTableID retrieves the table ID for the specified table. It uses the
+	// descriptor cache to perform lookups, falling back to the KV store when
+	// necessary.
+	getTableID(qname *parser.QualifiedName) (sqlbase.ID, error)
+
+	// NB: one can use getTableDescFromID() to retrieve a descriptor for
+	// a table from a transaction using its ID, assuming it was loaded
+	// in the transaction already.
+
+	// notifySchemaChange notifies that an outstanding schema change
+	// exists for the table.
+	notifySchemaChange(id sqlbase.ID, mutationID sqlbase.MutationID)
+
+	// getTableLease acquires a lease for the specified table. The lease will be
+	// released when the planner closes. Note that a shallow copy of the table
+	// descriptor is returned. It is safe to mutate fields of the returned
+	// descriptor, but the values those fields point to should not be modified.
+	getTableLease(qname *parser.QualifiedName) (sqlbase.TableDescriptor, error)
+
+	// getAliasedTableLease looks up the table descriptor for an alias table
+	// expression.
+	getAliasedTableLease(n parser.TableExpr) (*sqlbase.TableDescriptor, error)
+
+	// releaseLeases releases all leases currently held by the planner.
+	releaseLeases()
+
+	// writeTableDesc effectively writes a table descriptor to the
+	// database within the current planner transaction.
+	writeTableDesc(tableDesc *sqlbase.TableDescriptor) error
+}
+
+var _ SchemaAccessor = &planner{}
+
+// getTableDesc implements the SchemaAccessor interface.
 func (p *planner) getTableDesc(qname *parser.QualifiedName) (*sqlbase.TableDescriptor, error) {
 	if err := qname.NormalizeTableName(p.session.Database); err != nil {
 		return nil, err
@@ -85,45 +174,7 @@ func (p *planner) getTableDesc(qname *parser.QualifiedName) (*sqlbase.TableDescr
 	return &desc, nil
 }
 
-var errDescriptorNotFound = errors.New("descriptor not found")
-
-// get the table descriptor for the ID passed in using an existing txn.
-// returns an error if the descriptor doesn't exist or if it exists and is not
-// a table.
-func getTableDescFromID(txn *client.Txn, id sqlbase.ID) (*sqlbase.TableDescriptor, error) {
-	desc := &sqlbase.Descriptor{}
-	descKey := sqlbase.MakeDescMetadataKey(id)
-
-	if err := txn.GetProto(descKey, desc); err != nil {
-		return nil, err
-	}
-	table := desc.GetTable()
-	if table == nil {
-		return nil, errDescriptorNotFound
-	}
-	return table, nil
-}
-
-func getKeysForTableDescriptor(
-	tableDesc *sqlbase.TableDescriptor,
-) (zoneKey roachpb.Key, nameKey roachpb.Key, descKey roachpb.Key) {
-	zoneKey = sqlbase.MakeZoneKey(tableDesc.ID)
-	nameKey = sqlbase.MakeNameMetadataKey(tableDesc.ParentID, tableDesc.GetName())
-	descKey = sqlbase.MakeDescMetadataKey(tableDesc.ID)
-	return
-}
-
-func updateCommitBy(commitBy time.Time, lease *LeaseState) time.Time {
-	if commitBy.IsZero() || commitBy.After(lease.Expiration()) {
-		return lease.Expiration()
-	}
-	return commitBy
-}
-
-// getTableLease acquires a lease for the specified table. The lease will be
-// released when the planner closes. Note that a shallow copy of the table
-// descriptor is returned. It is safe to mutate fields of the returned
-// descriptor, but the values those fields point to should not be modified.
+// getTableLease implements the SchemaAccessor interface.
 func (p *planner) getTableLease(qname *parser.QualifiedName) (sqlbase.TableDescriptor, error) {
 	if err := qname.NormalizeTableName(p.session.Database); err != nil {
 		return sqlbase.TableDescriptor{}, err
@@ -175,9 +226,7 @@ func (p *planner) getTableLease(qname *parser.QualifiedName) (sqlbase.TableDescr
 	return lease.TableDescriptor, nil
 }
 
-// getTableID retrieves the table ID for the specified table. It uses the
-// descriptor cache to perform lookups, falling back to the KV store when
-// necessary.
+// getTableID implements the SchemaAccessor interface.
 func (p *planner) getTableID(qname *parser.QualifiedName) (sqlbase.ID, error) {
 	if err := qname.NormalizeTableName(p.session.Database); err != nil {
 		return 0, err
@@ -212,6 +261,7 @@ func (p *planner) getTableID(qname *parser.QualifiedName) (sqlbase.ID, error) {
 	return sqlbase.ID(gr.ValueInt()), nil
 }
 
+// getTableNames implements the SchemaAccessor interface.
 func (p *planner) getTableNames(dbDesc *sqlbase.DatabaseDescriptor) (parser.QualifiedNames, error) {
 	prefix := sqlbase.MakeNameMetadataKey(dbDesc.ID, "")
 	sr, err := p.txn.Scan(prefix, prefix.PrefixEnd(), 0)
@@ -236,4 +286,87 @@ func (p *planner) getTableNames(dbDesc *sqlbase.DatabaseDescriptor) (parser.Qual
 		qualifiedNames = append(qualifiedNames, qname)
 	}
 	return qualifiedNames, nil
+}
+
+// getAliasedTableLease implements the SchemaAccessor interface.
+func (p *planner) getAliasedTableLease(n parser.TableExpr) (*sqlbase.TableDescriptor, error) {
+	ate, ok := n.(*parser.AliasedTableExpr)
+	if !ok {
+		return nil, util.Errorf("TODO(pmattis): unsupported FROM: %s", n)
+	}
+	table, ok := ate.Expr.(*parser.QualifiedName)
+	if !ok {
+		return nil, util.Errorf("TODO(pmattis): unsupported FROM: %s", n)
+	}
+	desc, err := p.getTableLease(table)
+	if err != nil {
+		return nil, err
+	}
+	return &desc, nil
+}
+
+// notifySchemaChange implements the SchemaAccessor interface.
+func (p *planner) notifySchemaChange(id sqlbase.ID, mutationID sqlbase.MutationID) {
+	sc := SchemaChanger{
+		tableID:    id,
+		mutationID: mutationID,
+		nodeID:     p.evalCtx.NodeID,
+		cfg:        p.systemConfig,
+		leaseMgr:   p.leaseMgr,
+	}
+	p.session.TxnState.schemaChangers.queueSchemaChanger(sc)
+}
+
+// releaseLeases implements the SchemaAccessor interface.
+func (p *planner) releaseLeases() {
+	if p.leases != nil {
+		for _, lease := range p.leases {
+			if err := p.leaseMgr.Release(lease); err != nil {
+				log.Warning(err)
+			}
+		}
+		p.leases = nil
+	}
+}
+
+// writeTableDesc implements the SchemaAccessor interface.
+func (p *planner) writeTableDesc(tableDesc *sqlbase.TableDescriptor) error {
+	return p.txn.Put(sqlbase.MakeDescMetadataKey(tableDesc.GetID()),
+		sqlbase.WrapDescriptor(tableDesc))
+}
+
+// expandTableGlob implements the SchemaAccessor interface.
+func (p *planner) expandTableGlob(expr *parser.QualifiedName) (
+	parser.QualifiedNames, error) {
+	if len(expr.Indirect) == 0 {
+		return parser.QualifiedNames{expr}, nil
+	}
+
+	if err := expr.QualifyWithDatabase(p.session.Database); err != nil {
+		return nil, err
+	}
+	// We must have a single indirect: either .table or .*
+	if len(expr.Indirect) != 1 {
+		return nil, fmt.Errorf("invalid table glob: %s", expr)
+	}
+
+	switch expr.Indirect[0].(type) {
+	case parser.NameIndirection:
+		return parser.QualifiedNames{expr}, nil
+	case parser.StarIndirection:
+		dbDesc, err := p.getDatabaseDesc(string(expr.Base))
+		if err != nil {
+			return nil, err
+		}
+		if dbDesc == nil {
+			return nil, databaseDoesNotExistError(string(expr.Base))
+		}
+		tableNames, err := p.getTableNames(dbDesc)
+		if err != nil {
+			return nil, err
+		}
+		return tableNames, nil
+	default:
+		return nil, fmt.Errorf("invalid table glob: %s", expr)
+	}
 }
